ai_service/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
easyocr
huggingface_hub
langchain
langchain-huggingface
langchain-core
langfuse
python-dotenv
